Added files via upload
I was able to establish the beginnings of the Account & Reader classes. 
Only the AccountReader was implemented. 
I re-named and re-shuffled a couple things in the Store class. 
I changed where the EOF check is for looping through a file. 
I was able to create an Account pointer and create accounts via the data4customers text file
I'm not using any data structures yet to store the Accounts.